refactor(dependency-leases): extract lock+session store module (#677)
Split src/dependency-leases.ts (1239 -> 553-line facade) by
responsibility: dependency-leases-store.ts owns path/session-key
helpers, the interprocess session-lock machinery (lock owner records,
process-liveness probes, withInterprocessSessionLock/
withSessionMutationLock), filesystem-safety guards, and session
read/write state. Lock and store code are mutually recursive so they
stay in one module; the facade keeps the lease operations
(acquire/release/cleanup/withDependencyLeaseSession) and re-exports the
prior public surface. Single-describe test file unchanged.
